+++ source/extensions/isaacsim.replicator.behavior/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.1.15] - 2025-09-15
+### Changed
+- switched to isaacsim.test.utils for golden image comparison in tests
+
 ## [1.1.14] - 2025-07-17
 ### Changed
 - Updated windows golden images